feat: add variation of 'Add Manually' menu with manual value selection for all entries
