+++ applications/main/subghz/scenes/subghz_scene_config.h
@@ -12,9 +12,10 @@ ADD_SCENE(subghz, saved_menu, SavedMenu)
 ADD_SCENE(subghz, delete, Delete)
 ADD_SCENE(subghz, delete_success, DeleteSuccess)
 ADD_SCENE(subghz, set_type, SetType)
-ADD_SCENE(subghz, set_fix, SetFix)
+ADD_SCENE(subghz, set_key, SetKey)
+ADD_SCENE(subghz, set_serial, SetSerial)
 ADD_SCENE(subghz, set_button, SetButton)
-ADD_SCENE(subghz, set_cnt, SetCnt)
+ADD_SCENE(subghz, set_counter, SetCounter)
 ADD_SCENE(subghz, set_seed, SetSeed)
 ADD_SCENE(subghz, frequency_analyzer, FrequencyAnalyzer)
 ADD_SCENE(subghz, radio_settings, ExtModuleSettings)
