+++ applications/main/subghz/scenes/subghz_scene_save_name.c
@@ -103,12 +103,8 @@ bool subghz_scene_save_name_on_event(void* context, SceneManagerEvent event) {
                 furi_string_set(subghz->file_path, subghz->file_path_tmp);
             }
         }
-        if(scene_manager_has_previous_scene(subghz->scene_manager, SubGhzSceneSetSeed)) {
-            scene_manager_search_and_switch_to_previous_scene(
-                subghz->scene_manager, SubGhzSceneSetType);
-        } else {
-            scene_manager_previous_scene(subghz->scene_manager);
-        }
+
+        scene_manager_previous_scene(subghz->scene_manager);
 
         return true;
     } else if(event.type == SceneManagerEventTypeCustom) {
