+++ applications/main/subghz/scenes/subghz_scene_start.c
@@ -1,17 +1,8 @@
 #include "../subghz_i.h"
+#include "subghz_scene_start.h"
 #include <dolphin/dolphin.h>
 
 #include <lib/subghz/protocols/raw.h>
-
-enum SubmenuIndex {
-    SubmenuIndexRead = 10,
-    SubmenuIndexSaved,
-    SubmenuIndexAddManually,
-    SubmenuIndexFrequencyAnalyzer,
-    SubmenuIndexReadRAW,
-    SubmenuIndexExtSettings,
-    SubmenuIndexRadioSetting,
-};
 
 void subghz_scene_start_submenu_callback(void* context, uint32_t index) {
     SubGhz* subghz = context;
@@ -43,7 +34,7 @@ void subghz_scene_start_on_enter(void* context) {
     submenu_add_item(
         subghz->submenu,
         "Add Manually [Advanced]",
-        SubmenuIndexAddManually,
+        SubmenuIndexAddManuallyAdvanced,
         subghz_scene_start_submenu_callback,
         subghz);
     submenu_add_item(
@@ -92,6 +83,11 @@ bool subghz_scene_start_on_event(void* context, SceneManagerEvent event) {
         } else if(event.event == SubmenuIndexAddManually) {
             scene_manager_set_scene_state(
                 subghz->scene_manager, SubGhzSceneStart, SubmenuIndexAddManually);
+            scene_manager_next_scene(subghz->scene_manager, SubGhzSceneSetType);
+            return true;
+        } else if(event.event == SubmenuIndexAddManuallyAdvanced) {
+            scene_manager_set_scene_state(
+                subghz->scene_manager, SubGhzSceneStart, SubmenuIndexAddManuallyAdvanced);
             scene_manager_next_scene(subghz->scene_manager, SubGhzSceneSetType);
             return true;
         } else if(event.event == SubmenuIndexFrequencyAnalyzer) {
