+++ applications/main/subghz/scenes/subghz_scene_start.h
@@ -1,0 +1,12 @@
+#pragma once
+
+enum SubmenuIndex {
+    SubmenuIndexRead = 10,
+    SubmenuIndexSaved,
+    SubmenuIndexAddManually,
+    SubmenuIndexAddManuallyAdvanced,
+    SubmenuIndexFrequencyAnalyzer,
+    SubmenuIndexReadRAW,
+    SubmenuIndexExtSettings,
+    SubmenuIndexRadioSetting,
+};
