+++ applications/main/subghz/subghz.c
@@ -213,7 +213,7 @@ SubGhz* subghz_alloc(bool alloc_for_tx_only) {
 
     subghz_rx_key_state_set(subghz, SubGhzRxKeyStateIDLE);
 
-    subghz->secure_data = malloc(sizeof(SecureData));
+    subghz->gen_info = malloc(sizeof(GenInfo));
 
     if(!alloc_for_tx_only) {
         subghz->ignore_filter = subghz->last_settings->ignore_filter;
@@ -319,7 +319,7 @@ void subghz_free(SubGhz* subghz, bool alloc_for_tx_only) {
         subghz_history_free(subghz->history);
     }
 
-    free(subghz->secure_data);
+    free(subghz->gen_info);
 
     //TxRx
     subghz_txrx_free(subghz->txrx);
